Fix missing malloc nullptr check
This addresses the issue found with PVS-Studio and reported in
PR https://github.com/trailofbits/onesixtyone/pull/12

The `malloc` function is used in `read_communities` which is used
in `init_options` only if the user specifies a community file.

The solution proposed in PR #12 would skip all community strings for
which the allocation would fail (probably all of them after the first
one that fails).

It is probably better to exit the program and inform the user that
the allocation failed instead, and this is, what the changes in
this commit do.